Provide platform-dependant functions. Ensure scripts have headers.
Currently the system recongnition is based on the value of
product.system() (see
https://docs.python.org/3/library/platform.html#platform.system).

If the command cannot be executed due to some reason, Viesco will
provide information on the expected, current and targeted platforms.

For example:

[patch_name] .. Ignored \home\user\.vscode-oss. (expected: Linux, host: Windows, target: Windows)

+++ viesco.py
@@ -5,52 +5,41 @@ from argparse import ArgumentParser
 from importlib import import_module
 from math import ceil
 from pathlib import Path
-from sys import platform, stderr, stdout
+from platform import system as get_system_name
+from sys import stderr, stdout
 from typing import Any, Callable
 
 
 class ScriptWriter:
     def __init__(self, patcher: Patcher, patches: list[tuple[str, Any]], output_path: str | None):
         self.patcher = patcher
-        self.lines = []
-
         patcher.output = self
 
-        if output_path:
-            self.output_path = Path(output_path)
+        self.platform = ""
+        self.lines = []
 
-            if self.output_path.suffix in {".bat", ".cmd"}:
-                self._prepare = self._batch_prepare
-                self.comment = self._batch_comment
-                self.set_variable = self._batch_set_variable
-                self.remove_file = self._batch_remove_file
-            else:
-                msg = f"Unsupported output file extension {self.output_path.suffix}."
-                raise ValueError(msg)
-        else:
+        if not output_path:
             self._prepare = self._dummy
             self.comment = self._dummy
             self.set_variable = self._dummy
             self.remove_file = self._dummy
+            return
 
-        self._prepare(patches)
-        self.set_variable("VSC_PATH", patcher.install_path)
+        self.output_path = Path(output_path)
+        if self.output_path.suffix in {".bat", ".cmd"}:
+            self.platform = "Windows"
+            self._prepare = self._batch_prepare
+            self.comment = self._batch_comment
+            self.set_variable = self._batch_set_variable
+            self.remove_file = self._batch_remove_file
+        else:
+            msg = f"Unsupported output file extension {self.output_path.suffix}."
+            raise ValueError(msg)
 
-    def write(self):
-        if self.lines:
-            self.output_path.write_text("\n".join(self.lines))
-
-    def _dummy(self, *_, **__):
-        pass
-
-    def _batch_prepare(self, patches: list[tuple[str, Any]]):
-        name = self.patcher.installed_name
-        version = self.patcher.installed_version
-
-        self.lines.append("@echo off")
-        self._batch_comment(
+        self._prepare()
+        self.comment(
             f"This script was created AUTOMATICALLY using Viesco v{Patcher.VERSION}",
-            f"for {name} v{version} on {platform}.",
+            f"for {patcher.host_product} v{patcher.host_version} on {patcher.host_platform}.",
             "",
             "Applied patches:",
             f"  {','.join(name for name, _ in patches)}",
@@ -67,6 +56,17 @@ class ScriptWriter:
             "warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR ",
             "PURPOSE.",
         )
+        self.set_variable("VSC_PATH", patcher.install_path)
+
+    def write(self):
+        if self.lines:
+            self.output_path.write_text("\n".join(self.lines))
+
+    def _dummy(self, *_, **__):
+        pass
+
+    def _batch_prepare(self):
+        self.lines.append("@echo off")
 
     def _batch_comment(self, *lines: str):
         self.lines.extend(f"rem {line}" for line in lines)
@@ -75,7 +75,9 @@ class ScriptWriter:
         self.lines.append(f'set "{name}={value}"')
 
     def _batch_remove_file(self, path: Path):
-        path = path.relative_to(self.patcher.install_path)
+        if self.patcher.install_path in path.parents:
+            path = path.relative_to(self.patcher.install_path)
+
         self.lines.append(f"echo :: Deleting {path}...")
         self.lines.append(f"del /F /Q %VSC_PATH%\\{path}")
 
@@ -85,22 +87,27 @@ class Patcher:
 
     def __init__(self, *, dry_run: bool = False):
         self.dry_run = dry_run
-        self.output: ScriptWriter = None  # pyright: ignore[reportAttributeAccessIssue]
-        self.install_path: Path = None  # pyright: ignore[reportAttributeAccessIssue]
+        self.output: ScriptWriter = None
+        self.install_path: Path = None
 
         self._skip_patch = False
-        self.current_patch = ""
+        self._current_patch = ""
 
-        self.product_info: dict = None  # pyright: ignore[reportAttributeAccessIssue]
-        self.package_info: dict = None  # pyright: ignore[reportAttributeAccessIssue]
+        self._product_info: dict = None
+        self._package_info: dict = None
 
-    @property
-    def installed_version(self) -> str:
-        return self.package_info["version"]
+        self.host_platform = get_system_name()
+        self.host_product = ""
+        self.host_version = ""
 
-    @property
-    def installed_name(self) -> str:
-        return self.product_info["nameShort"]
+    def _ask_to_skip_patch(self, *args):
+        self.print(*args, level="warning")
+        try:
+            self._skip_patch = (
+                input(f"[{self._current_patch}] Skip the patch? [Y/n]: ").lower() != "n"
+            )
+        except KeyboardInterrupt:
+            exit(1)
 
     def load_install_path(self, install_path: Path) -> bool:
         product_path = install_path / "resources/app/product.json"
@@ -110,47 +117,35 @@ class Patcher:
             return False
 
         with product_path.open() as f:
-            self.product_info = json.load(f)
+            self._product_info = json.load(f)
         with package_path.open() as f:
-            self.package_info = json.load(f)
+            self._package_info = json.load(f)
+
+        self.host_product = self._product_info["nameShort"]
+        self.host_version = self._package_info["version"]
 
         self.install_path = install_path
         return True
 
-    def _find(self, array: tuple, item: Any) -> int:
-        try:
-            return array.index(item)
-        except ValueError:
-            return -1
-
-    def _ask_to_skip_patch(self, *args):
-        self.print(*args, level="warning")
-        try:
-            self._skip_patch = (
-                input(f"[{self.current_patch}] Skip the patch? [Y/n]: ").lower() != "n"
-            )
-        except KeyboardInterrupt:
-            exit(1)
-
     def validate_patch(self, name: str, validator: Callable[[Patcher], None]) -> bool:
-        self.current_patch = name
+        self._current_patch = name
         validator(self)
         return not self._skip_patch
 
     def check_product_name(self, *supported: str):
-        if self.installed_name not in supported:
-            self._ask_to_skip_patch(f"'{self.installed_name}' is not supported by the patch.")
+        if self.host_product not in supported:
+            self._ask_to_skip_patch(f"'{self.host_product}' is not supported by the patch.")
 
     def check_version(self, minimal_str: str):
         minimal = tuple(map(int, minimal_str.split(".")))
-        current: tuple[int, ...] = tuple(map(int, self.installed_version.split(".")))
+        current: tuple[int, ...] = tuple(map(int, self.host_version.split(".")))
         if not (
             current[0] >= minimal[0]
             and (len(minimal) < 2 or current[1] >= minimal[1])
             and (len(minimal) < 3 or current[2] >= minimal[2])
         ):
             self._ask_to_skip_patch(
-                f"{self.installed_name} v{self.installed_version}",
+                f"{self.host_product} v{self.host_version}",
                 f"is not supported by the patch (minimal: v{minimal_str}).",
             )
 
@@ -161,7 +156,7 @@ class Patcher:
 
         self.print_items_with_index(labels)
 
-        prompt = f"[{self.current_patch}] {prompt} (comma-separated) > "
+        prompt = f"[{self._current_patch}] {prompt} (comma-separated) > "
         while True:
             try:
                 selected = input(prompt)
@@ -171,7 +166,11 @@ class Patcher:
             invalid = []
             selected = selected.split(",")
             for sel_idx, sel_item in enumerate(selected):
-                label_idx = self._find(labels, sel_item)
+                try:
+                    label_idx = labels.index(sel_item)
+                except ValueError:
+                    label_idx = -1
+
                 if label_idx < 0:
                     try:
                         label_idx = int(sel_item) - 1
@@ -225,8 +224,9 @@ class Patcher:
             )
 
     def print_patch_name(self, name: str):
+        self._current_patch = ""
         self.print(f"Starting '{name}'...")
-        self.current_patch = name
+        self._current_patch = name
 
         # Maybe it should be available as a ScriptWriter method.
         # Though it is needed only here, so hardcoding it is fine for now.
@@ -246,17 +246,34 @@ class Patcher:
         else:
             level = "::"
 
-        if self.current_patch:
-            print(f"[{self.current_patch}]", level, *args, file=fd, **kwargs)
+        if self._current_patch:
+            print(f"[{self._current_patch}]", level, *args, file=fd, **kwargs)
         else:
             print(level, *args, file=fd, **kwargs)
 
-    def remove(self, path: Path):
-        if not self.dry_run:
-            path.unlink(missing_ok=True)
+    def remove(self, path: Path | str, platform: str = ""):
+        path = Path(path)
+        is_targeting_host = not self.dry_run and (not platform or platform == self.host_platform)
+        is_targeting_output = self.output.lines and (
+            not platform or platform == self.output.platform
+        )
 
-        self.output.remove_file(path)
-        self.print(f"Removed {path}.", level="info")
+        if is_targeting_host:
+            path.unlink(missing_ok=True)
+            self.print(f"Removed {path}.", level="info")
+
+        if is_targeting_output:
+            self.output.remove_file(path)
+            if not is_targeting_host:
+                self.print(f"The script will remove {path}.", level="info")
+        elif not is_targeting_host:
+            self.print(
+                f"Ignored {path}.",
+                f"(expected: {platform or '<any>'},",
+                f"host: {self.host_platform or '<unknown>'},",
+                f"target: {self.output.platform or '<none>'})",
+                level="debug",
+            )
 
 
 if __name__ == "__main__":
@@ -282,10 +299,13 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     patcher = Patcher(dry_run=args.dry_run)
-    patcher.print(f"Received args: {args}", level="debug")
+    patcher.print(f"Received arguments: {args}", level="debug")
 
     if patcher.dry_run:
-        patcher.print("This is a dry run. No changes will be made.", level="warning")
+        patcher.print(
+            "This is a dry run. No changes will be made to the existing installation.",
+            level="warning",
+        )
     if not patcher.load_install_path(Path(args.install)):
         patcher.print(f"'{args.install}' is not a valid path to VSCodium.", level="warning")
         exit(1)
